fix binding extension missing method in development
