+++ config/environment.rb
@@ -3,6 +3,7 @@ require_relative 'application'
 
 require 'errors'
 require 'rescue_from_unless_local'
+require 'v0_bindings_extensions'
 
 # Initialize the Rails application.
 Rails.application.initialize!
